+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - PR [#117](https://github.com/marinasundstrom/CheckedExceptions/pull/#117) Handle redundant typed catch #114
 
+### Fixed
+
+- PR [#121](https://github.com/marinasundstrom/CheckedExceptions/pull/#121) Remove leading trivia before adding ThrowsAttribute
+
 ## [1.5.2] - 2025-07-26
 
 ## Added